Refactor: separate components related `TaskList`
